requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
redis==5.0.7
psycopg2-binary==2.9.9
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0
python-dotenv==1.0.1
numpy==2.0.1
